+++ src/webpubsub/azext_webpubsub/_params.py
@@ -32,11 +32,11 @@ def load_arguments(self, _):
         c.argument('webpubsub_name', webpubsub_name_type, options_list=['--name', '-n'])
 
     with self.argument_context('webpubsub create') as c:
-        c.argument('sku', help='The sku name of the webpubsub service. Allowed values: Free_F1, Standard_S1')
+        c.argument('sku', help='The sku name of the webpubsub service. Allowed values: Free_F1, Standard_S1, Premium_P1')
         c.argument('unit_count', help='The number of webpubsub service unit count', type=int)
 
     with self.argument_context('webpubsub update') as c:
-        c.argument('sku', help='The sku name of the webpubsub service. Allowed values: Free_F1, Standard_S1')
+        c.argument('sku', help='The sku name of the webpubsub service. Allowed values: Free_F1, Standard_S1, Premium_P1')
         c.argument('unit_count', help='The number of webpubsub service unit count', type=int)
 
     with self.argument_context('webpubsub key regenerate') as c: