+++ src/automation/HISTORY.rst
@@ -6,3 +6,8 @@ Release History
 0.1.0
 ++++++
 * Initial release.
+
+0.1.2
+++++++
+* Added Hybrid Runbook Worker Group and Hybrid Runbook Workers related commands.
+