+++ src/rdbms-connect/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+1.0.3
+++++++
++ Add support to read UTF-8 files with and without BOM
+
 1.0.2
 ++++++
 + Use compatible release for `setproctitle` to support Python 3.10